pyproject.toml
[tool.poetry]
name = "amocrm_integration"
version = "0.1.0"
description = ""
authors = ["Pavel Lyulchak"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
psycopg2-binary = "^2.9.10"
requests = "^2.32.3"
pydantic-settings = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
